Añadiendo permisos en post y correcciones en tabla

+++ app/Http/Controllers/PostController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Models\Post;
 use App\Models\Category;
 use Illuminate\Http\Response;
+use Illuminate\Support\Facades\Gate;
 use App\Http\Requests\Post\StoreRequest;
 use App\Http\Requests\Post\UpdateRequest;
 
@@ -17,17 +18,29 @@ class PostController extends Controller
      */
     public function index()
     {
-        return $this->successResponse(Post::with('categories')->get());
+        if (Gate::denies('post_index')) {
+            return $this->errorResponse(
+                '403 Forbidden', 
+                Response::HTTP_FORBIDDEN);
+        }
+        $posts = Post::with('categories')->get();
+        return $this->successResponse($posts);
     }
 
-     /**
+    /**
      * Show the form for creating a new resource.
      *
      * @return \Illuminate\Http\Response
      */
     public function create()
     {
-        return $this->successResponse(Category::all());
+        if (Gate::denies('post_add')) {
+            return $this->errorResponse(
+                '403 Forbidden', 
+                Response::HTTP_FORBIDDEN);
+        }
+        $categories = Category::all();
+        return $this->successResponse($categories);
     }
 
     /**
@@ -38,6 +51,11 @@ class PostController extends Controller
      */
     public function store(StoreRequest $request)
     {
+        if (Gate::denies('post_add')) {
+            return $this->errorResponse(
+                '403 Forbidden', 
+                Response::HTTP_FORBIDDEN);
+        }
         $inputs = $request->validated();
 
         $post = Post::create($inputs);
@@ -59,7 +77,12 @@ class PostController extends Controller
      */
     public function show(Post $post)
     {
-        return $this->successResponse($post);
+        if (Gate::denies('post_index')) {
+            return $this->errorResponse(
+                '403 Forbidden', 
+                Response::HTTP_FORBIDDEN);
+        }
+        return $this->successResponse($post->load('categories'));
     }
 
     /**
@@ -70,7 +93,12 @@ class PostController extends Controller
      */
     public function edit(Post $post)
     {
-        $data['post']       = Post::where('id', $post->id)->with('categories')->get();
+        if (Gate::denies('post_edit')) {
+            return $this->errorResponse(
+                '403 Forbidden', 
+                Response::HTTP_FORBIDDEN);
+        }
+        $data['post']       = $post->load('categories');
         $data['categories'] = Category::all();
 
         return $this->successResponse($data);
@@ -85,6 +113,11 @@ class PostController extends Controller
      */
     public function update(UpdateRequest $request, Post $post)
     {
+        if (Gate::denies('post_edit')) {
+            return $this->errorResponse(
+                '403 Forbidden', 
+                Response::HTTP_FORBIDDEN);
+        }
         $post = Post::find($post)->first();
 
         $inputs = $request->validated();
@@ -107,6 +140,11 @@ class PostController extends Controller
      */
     public function destroy(Post $post)
     {
+        if (Gate::denies('post_delete')) {
+            return $this->errorResponse(
+                '403 Forbidden', 
+                Response::HTTP_FORBIDDEN);
+        }
         Post::destroy($post->id);
 
         return $this->successResponse('', 'Post deleted successfully.', Response::HTTP_OK);
